### 2022-10-26 #4884 @kyleniemeyer
@openjournals/dev can someone help figure out why the paper isn't generating? I can see the `paper.md` file on the `paper.md` branch, which I've now set. I can't parse the error message that is being given.
