### 2022-10-26 e805c94
add new tuts to doctree and introduce crossref labels
