+++ docs/source/tutorials.rst
@@ -5,4 +5,5 @@ Tutorials
 .. toctree::
 
    tutorials/introduction.rst
-
+   tutorials/benchmark.rst
+   tutorials/parallel.rst
