+++ docs/source/tutorials/benchmark.rst
@@ -1,0 +1,181 @@
+========================
+Benchmarking GGCE
+========================
+
+In this tutorial, we will show GGCE in operation. Specifically, we will
+demonstrate how the objects you learned about in the previous section --
+``Model``, ``System``, and ``Solver`` -- can be used to produce a spectral
+function :math:`A(k,\omega)`.
+
+To benchmark GGCE, we will compare our result against a prior numerical study
+(`Berciu & Fehske, PRB 82, 085116 (2010) <http://link.aps.org/abstract/PRB/v82/e085116>`__).
+In that study, the authors employed the momentum average approximation (MA) --
+a technique from which GGCE directly descends -- to calculate the spectrum of
+the Edwards Fermion Boson model we introduced in the previous tutorial.
+
+As a first sanity check, we will compare GGCE to its ancestor.
+
+Step 1: prepare the calculation
+-------------------------------
+
+In the reference paper linked above, a single-particle Green's function and
+spectrum are calculated for the Edwards Fermion Boson coupling Hamiltonian,
+
+.. math::
+
+    H = -t \sum_{\langle i, j \rangle} c_i^\dagger c_i + \Omega \sum_i b_i^\dagger b_i - g \sum_{\langle i, j \rangle} c_i^\dagger c_j \left( b_j^\dagger + b_i \right)
+
+For the purposes of this demonstration, it is not important what this
+model represents: as long as it can be expressed in terms of
+creation/annihilation operators :math:`b_j^\dagger, b_j, c_j^\dagger, c_j`,
+GGCE can give us its spectrum (but see the note below for some physical
+intuition for the model).
+
+.. note ::
+
+  An intuitive picture for this model of coupling is the following: every
+  time a fermion hops to a site `j` with no bosons, it leaves behind a
+  boson excitation created by :math:`b^\dagger_j`. On the other hand,
+  if the fermion hops to a site where a boson already exists, the boson
+  gets destroyed by :math:`b_i`.
+
+  This might seem like a strange model at first glance. However, taking
+  fermions to be holes and the bosonic excitations to be "spin-flips",
+  or magnons, this model allows one to "mimic the motion" of a charge
+  carrier "through an antiferromagnetically ordered spin background",
+  such as thought to be relevant in the context of cuprate CuO layers.
+  See Sec. II in
+  `Berciu & Fehske, PRB 82, 085116 (2010) <http://link.aps.org/abstract/PRB/v82/e085116>`__
+  for more details.
+
+The ``Model`` can be initialized as before
+
+.. code-block:: python
+
+    from ggce import Model
+    model = Model.from_parameters(hopping=0.1)
+    model.add_(
+        "EdwardsFermionBoson",
+        phonon_frequency=1.25,
+        phonon_extent=3,
+        phonon_number=9,
+        dimensionless_coupling_strength=2.5
+    )
+
+Here we specifically picked model parameters corresponding to those used
+in Fig. 5 (top panel) from the reference paper.
+
+.. note::
+
+  While the ``phonon_number`` is not set explicitly in the reference,
+  in MA it is understood that it is varied until physical results,
+  such as spectra, are converged relative to it. Here we found the value
+  ``phonon_number = 9`` resulted in appropriate convergence. In general,
+  a full convergence study in all variational parameters should be conducted
+  to make sure convergence is reached.
+
+Creating the ``System``, we optionally choose to pass the option
+``autoprime=False``. This defers the building of the system of
+equations within ``System`` until we instantiate a ``Solver``.
+
+.. code-block:: python
+
+    from ggce import System
+    system = System(model, autoprime=False)
+
+.. note::
+
+  In a large variational calculation, the creation -- or "priming" --
+  of the ``System`` object can take a long time.
[…]
