+++ docs/source/tutorials/parallel.rst
@@ -1,0 +1,5 @@
+.. _parallel:
+
+===================================
+Running GGCE in parallel (no PETSc)
+===================================
