### 2022-10-26 0882411
Create CITATION.cff
