### 2022-10-26 276cd61
Delete CITATION.cff
Added citation information in README.
