+++ docs/source/tutorials.rst
@@ -7,3 +7,5 @@ Tutorials
    tutorials/introduction.rst
    tutorials/benchmark.rst
    tutorials/parallel.rst
+   tutorials/petsc_intro.rst
+   tutorials/petsc_advanced.rst
