### 2022-10-27 3820917
typo fixes in benchmark tutorial
