### 2022-10-27 2b806fb
ne tutorial on parallel tecchniques in ggce
