+++ docs/source/tutorials/petsc_intro.rst
@@ -1,0 +1,5 @@
+.. _petsc-intro:
+
+====================
+Using PETSc for GGCE
+====================
