+++ docs/source/tutorials/finite_t.rst
@@ -1,0 +1,5 @@
+.. _finite-t:
+
+==============================
+Finite-temprature calculations
+==============================
