### 2022-10-27 #4884 @xuanxu
The problem is caused by the branch being named `paper.md`. The bot looks for a `paper.md` file in the repository to compile it, and the first one is finding now is the internal git file referencing that branch. Renaming the branch should fix the problem.
