### 2022-10-27 e521005
Update README.rst
Add JOSS badge
