+++ docs/source/tutorials.rst
@@ -9,3 +9,4 @@ Tutorials
    tutorials/parallel.rst
    tutorials/petsc_intro.rst
    tutorials/petsc_advanced.rst
+   tutorials/finite_t.rst
