+++ docs/source/tutorials/parallel.rst
@@ -1,8 +1,8 @@
 .. _parallel:
 
-===================================
-Running GGCE in parallel (no PETSc)
-===================================
+========================
+Running GGCE in parallel
+========================
 
 Available parallel schemes
 --------------------------
