### 2022-11-08 1a0b102
added petsc intro tutorial with images
