+++ docs/source/tutorials/petsc_advanced.rst
@@ -1,5 +1,179 @@
 .. _petsc-advanced:
 
-==================================
-GGCE with PETSc: advanced features
-==================================
+=======================================
+GGCE-PETSc interface: advanced features
+=======================================
+
+In this tutorial we go over two advanced features enabled by the GGCE-PETSc interface:
+the double-parallel scheme and the basis-matrix separation, which together unlock
+large-scale phonon configuration calculations.
+
+Double-parallel scheme
+----------------------
+
+In Sec. :ref:`parallel`, we alluded to the double-parallel scheme of running GGCE.
+To recap: in the double-parallel scheme, the calculation is MPI-parallelized BOTH across different
+:math:`(k,\omega)` points, AND within the solution of the matrix equation at each individual
+point. To enable and control this feature when running GGCE-PETSc, we use the ``brigade_size``
+keyword when instantiating the Solver.
+
+.. code-block:: python
+
+  COMM = MPI.COMM_WORLD
+  solver = MassSolverMUMPS(system=system, mpi_comm=COMM, brigade_size=1)
+
+This is the only syntactic change required to enable this scheme.
+
+By a "brigade", we mean a group of MPI processes
+that are collectively working on a single :math:`(k,\omega)` point. The keyword
+``brigade_size`` is used in the code to split all MPI processes into brigades of
+**equal** size ``brigade_size``. Each brigade works collectively on the matrix for a single
+:math:`(k,\omega)` point, and together the brigades cover the entire :math:`(k,\omega)`
+space.
+
+By default, ``brigade_size=WORLD_SIZE``, meaning that there is only one brigade,
+and the Solver proceeds sequentially through all :math:`(k,\omega)` points whilst
+maximally distributing the matrix of equations across all available MPI processes.
+The main limitation for this keyword is that the number of MPI processes should be
+evenly divisible by ``brigade_size``, otherwise an Error is raised and the calculation
+is terminated.
+
+.. note ::
+
+  Due to the specifics of the PETSc
+  API, the number of "jobs" fed into ``Solver.greens_function(k, w, eta)`` --
+  that is, ``len(k)*len(w)`` -- is required to be evenly divisible by the number of brigades
+  (otherwise the PETSc solvers hang indefinitely). The Solver automatically extends (pads)
+  the :math:`k, \omega` arrays with the minimum needed number of extra points to
+  prevent this. The padding is done to the :math:`omega` array by default, unless it
+  is of length 1: in that case, the :math:`k` array is padded. The padding is always
+  done outside the range specified by the user, so the input :math:`k,\omega` arrays
+  are NOT redefined but merely extended, so that the user still has perfect
+  control over where the calculations are being carried out.
+
+  All calculations, including over the padded points, are returned by
+  ``greens_function``. This is why sometimes the output array of ``greens_function()``
+  may have slightly larger dimensions than the original input array. This is of
+  course easily fixed by merely indexing the output array up to the lengths of the
+  original :math:`k,\omega` arrays
+
+  .. code-block:: python
+
+    result = solver.greens_function(k, w, eta)
+    result_notpadded = result[:len(k),:len(w)]
+
+The optimal choice of ``brigade_size`` is ultimately up to the user. For small
+systems where the matrix can be easily stored and solved by a single CPU, the
+optimal speed-up is accomplished with maximum splitting, i.e. ``brigade_size=1``.
+When dealing with larger matrices (say of dimension more than 100,000), optimal
+speed-up will have to be determined by experimenting. In practice, ``brigade_size``
+should be just enough for the matrix solution to be possible within MUMPS in terms of
+memory and time, so that the largest number of independent brigades can be formed.
+
+For the calculation in the previous section, we see that the sizes of matrices
+being solved are quite small (merely hundreds of equations). This allows us to set
[…]
