### 2022-11-11 #4884 @jarvist
OK, @x94carbone , do you have any suggestions for Reviewers?
I've invited @Neutrino155, a PhD student I supervise, to review, as your package + paper strongly overlaps with their expertise, and I suspect it might be difficult to find two fully external Polaron reviewers.
