### 2022-11-11 #4884 @x94carbone
@jarvist so I searched the spreadsheet before and found 13 people with "condensed matter" in their expertise. Would you like me to try and select a few of those who it seems overlaps with _theoretical_ condensed matter and suggest them here?
