### 2022-11-12 71b75e2
Update logger docs
