### 2022-11-14 #4884 @x94carbone
@jarvist perhaps we could start with (in order, people at the top of this list are at the bottom of the Google Sheets doc):
- merrygoat
- santoshkumarradha
- Evan1415
- vyasr

This group has some combination of expertise in condensed matter, HPC and Python
