+++ ggce/__init__.py
@@ -2,6 +2,7 @@ from .logger import logger  # noqa
 from .model import Model  # noqa
 from .engine.system import System  # noqa
 from .executors.solvers import SparseSolver, DenseSolver  # noqa
+from .executors.petsc4py.solvers import MassSolverMUMPS  # noqa
 
 
 # DO NOT CHANGE ANYTHING BELOW THIS
