### 2022-11-25 5d7872a
simplify petsc import with try-except
