+++ ggce/__init__.py
@@ -2,8 +2,10 @@ from .logger import logger  # noqa
 from .model import Model  # noqa
 from .engine.system import System  # noqa
 from .executors.solvers import SparseSolver, DenseSolver  # noqa
-from .executors.petsc4py.solvers import MassSolverMUMPS  # noqa
-
+try:
+    from .executors.petsc4py.solvers import MassSolverMUMPS  # noqa
+except ImportError:
+    pass
 
 # DO NOT CHANGE ANYTHING BELOW THIS
 __version__ = ...  # semantic-version-placeholder
