### 2022-11-25 103915b
add --runslow flag to CI tests
