### 2022-11-29 b435782
remove --runslow flag as test runs out of github time
