### 2022-12-04 #4884 @x94carbone
@jarvist just bumping this! Any updates?

Also, @kyleniemeyer I plan on changing my GitHub username shortly. Will this cause any issues for the review/eventual JOSS manuscript? Figured I'd mention this before I did change it.

Thanks!
