### 2022-12-05 #4884 @kyleniemeyer
@x94carbone changing your username *shouldn't* cause any issues, but I guess we'll find out! @openjournals/dev any thoughts here?
