### 2022-12-05 #4884 @xuanxu
I also think there's should not cause any problem. As long as the right username and url (the repo's url will change with the username) are updated in the issue body, everything should work as expected.
