+++ legacy/examples/scripts/fit_lorentz_scipy.py
@@ -1,213 +1,0 @@
-import numpy as np
-from scipy.optimize import curve_fit
-
-# uncomment if want to visualize PETSc speed benchmark
-# import matplotlib as mpl
-import matplotlib.pyplot as plt
-
-import sys
-import os
-import shutil
-import time
-
-script_dir = os.path.dirname(os.path.realpath(__file__))
-ggce_head_dir = (
-    f"/mnt/c/Users/bobst/Documents/University_of_British_Columbia/"
-    f"Physics/Mona_Berciu/Generalized_Green_function_cluster_expan"
-    f"sion/new_lorentzfit"
-)
-data_dir = (
-    f"/mnt/c/Users/bobst/Documents/University_of_British_Columbia/"
-    f"Physics/Mona_Berciu/Generalized_Green_function_cluster_expan"
-    f"sion/fin_T_hol"
-)
-sys.path.append(ggce_head_dir)
-import ggce
-
-# load the mpi communicator
-from mpi4py import MPI
-
-# load the Model and Executor classes to store parameters and manage the calculation
-from ggce.model import Model  # noqa: E402
-from ggce.executors.petsc4py.parallel import ParallelSparseExecutorMUMPS
-from ggce.executors.serial import SerialDenseExecutor
-from ggce.utils.utils import lorentzian, peak_location_and_weight_scipy
-
-server = f"lisa"
-Mval = 3
-Nval = 9
-Mval2 = 3
-Nval2 = 5
-lam, om, t = 1, 0.5, sys.argv[1]
-eta = 0.05
-date = sys.argv[2]
-aux = f"polaron_dispersion"
-year = f"22"
-krange = [
-    "0.000",
-    "0.008",
-    "0.017",
-    "0.025",
-    "0.033",
-    "0.041",
-    "0.050",
-    "0.058",
-    "0.066",
-    "0.074",
-    "0.083",
-    "0.091",
-    "0.099",
-    "0.107",
-    "0.116",
-    "0.124",
-    "0.132",
-    "0.141",
-    "0.149",
-    "0.157",
-]
-
-results_dir = os.path.join(
-    data_dir,
-    f"lam_{lam}",
-    f"om_{om}",
-    f"t_{t}",
-    aux,
-    f"{Mval}{Nval}_{Mval2}{Nval2}_{server}_{date}_{year}",
-    f"pol_dis_HfT_{Mval}{Nval}_{Mval2}{Nval2}",
-)
-
-energy_k = np.zeros(len(krange))
-energy_k_err = np.zeros(len(krange))
-eta_k = np.zeros(len(krange))
-
-fig0, ax0 = plt.subplots()
-for ii, k in enumerate(krange):
-    result_file = os.path.join(results_dir, f"all_results_k_{k}.txt")
-    wrange, Arange = np.loadtxt(result_file, unpack=True, skiprows=1)
-
-    # restrict wrange to last 10 points near peak
-    wrange = wrange[-10:]
-    Arange = Arange[-10:]
-
-    fit_params, error = peak_location_and_weight_scipy(wrange, Arange, eta)
-
-    wrange_pred = np.linspace(wrange[0], wrange[-1], 100)
-    Arange_pred = lorentzian(wrange_pred, *fit_params)
-
-    if ii in [0, 10, 19]:
-        ax0.plot(wrange_pred, Arange_pred)
-        ax0.scatter(wrange, Arange)
-
-    energy_k[ii] = fit_params[0]
-    energy_k_err[ii] = error[0, 0]
-    eta_k[ii] = fit_params[2]
-
-# show together the plots of the lorentz fits
-plt.show()
-
-result_file = os.path.join(results_dir, f"polaron_dispersion.txt")
[…]
