+++ legacy/benchmarks/reliability_petsc.py
@@ -1,171 +1,0 @@
-############ Stress-testing PETSc ##########
-"""
-In this script we push PETSc to the limit. We will calculate entire spectral
-slices, tracking both time to evaluate as well as saving the actual spectra.
-We are looking to evalaute convergence wiht (M,N), as well as time to calculate
-"""
-########### ################
-
-# before we import numpy, set the number of threads for numpy multithreading
-# using the OMP_NUM_THREADS environment variable
-import numpy as np
-
-# uncomment if want to visualize PETSc speed benchmark
-# import matplotlib as mpl
-import matplotlib.pyplot as plt
-
-import sys
-import os
-import shutil
-
-script_dir = os.path.dirname(os.path.realpath(__file__))
-ggce_head_dir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
-
-try:
-    import ggce
-except ModuleNotFoundError:
-    sys.path.append(ggce_head_dir)
-    import ggce
-
-# load the mpi communicator
-from mpi4py import MPI
-
-# load the Model and Executor classes to store parameters and manage the calculation
-from ggce.model import Model  # noqa: E402
-from ggce.executors.petsc4py.parallel import ParallelSparseExecutorMUMPS
-
-# run PETSc benchmark for cloud size up to cloud_max with up to bosons_max bosons
-cloud_min = 7
-cloud_max = 12
-bosons_min = 6
-bosons_max = 15
-
-# fix parameter values same as ground truth from the https://journals.aps.org/prb/
-# abstract/10.1103/PhysRevB.82.085116, figure 5 center for comparison
-Omega = 1.25
-coupling = 2.5
-hopping = 0.1
-eta = 0.005
-
-# load the "grouth truth" data from https://journals.aps.org/prb/
-# abstract/10.1103/PhysRevB.82.085116, figure 5 center for comparison
-literature_data = np.loadtxt(
-    os.path.join(ggce_head_dir, "examples", "000_example_A.txt")
-)
-
-# define w and k grid for the calculation
-wgrid = np.linspace(-5.5, -2.5, 3)
-k = 0.5 * np.pi
-
-# loop over the range of allowed N_bosons to full spectrum slice (k,wgrid)
-# start with defining an MPI communicator
-COMM = MPI.COMM_WORLD
-
-# start by creating a folder with the text and image output in run directory
-# only head node make directories
-results_dir = os.path.join(script_dir, "stress_test_results")
-spectra_vis_path = os.path.join(
-    script_dir, "stress_test_results", "spectra_visuals"
-)
-text_data_path = os.path.join(script_dir, "stress_test_results", "text_data")
-
-if COMM.Get_rank() == 0:
-    if not os.path.exists(results_dir):
-        os.mkdir(results_dir)
-        # now create nested directories for text and image respectively
-        os.mkdir(spectra_vis_path)
-        os.mkdir(text_data_path)
-    else:
-        shutil.rmtree(results_dir)
-        os.mkdir(results_dir)
-        os.mkdir(spectra_vis_path)
-        os.mkdir(text_data_path)
-
-# create an array to store time to finish the calculation
-parallel_times = np.zeros((bosons_max, cloud_max))
-
-# start the loop over all allowed M, N
-for m in range(cloud_min, cloud_max + 1):
-    for n in range(bosons_min, bosons_max + 1):
-        for w in wgrid:
-            # initialize the model with pre-determined parameters
-            model = Model()
-            model.set_parameters(hopping=hopping)
-            model.add_coupling(
-                "EdwardsFermionBoson",
-                Omega=Omega,
-                M=m,
-                N=n,
-                dimensionless_coupling=coupling,
-            )
-
[…]
