### 2023-01-17 4d51fe0
hotfix to make sure petsc KSP solver is created on appropriate brigadier MPI_COMM
