### 2023-01-29 #5115 @editorialbot
[REVIEW]: The Generalized Green's function Cluster Expansion: A Python package for simulating polarons
**Submitting author:** @matthewcarbone (<a href="http://orcid.org/0000-0002-5181-9513">Matthew Carbone</a>)
**Repository:** https://github.com/matthewcarbone/GGCE
**Branch with paper.md** (empty if default branch): joss-manuscript
**Version:** v0.1.4
