### 2023-01-29 #5115 @jarvist
Thank you Stefano @st3r4g Brad @Neutrino155 for agreeing to review! The key instructions and links to guidance are in the first piece of boilerplate written by @editorialbot above ^^^^.
I also find the list of commands for @editorialbot very helpful. https://joss.readthedocs.io/en/latest/editorial_bot.html
