### 2023-01-30 93af781
fix link to installation docs in README.rst
