+++ pyproject.toml
@@ -30,7 +30,7 @@ classifiers = [
     "Intended Audience :: Science/Research",
 ]
 dependencies = [
-    "numpy",
+    "numpy==1.23",
     "scipy",
     "loguru",
     "monty",
