+++ ggce/_tests/petsc/test_petsc_sysgen.py
@@ -314,7 +314,6 @@ EFB_Figure6_k2_params = {
     ],
 )
 def test_prb_82_085116_2010(p):
-
     COMM = MPI.COMM_WORLD
 
     gt = p["gt"]
