+++ ggce/_tests/mpi/test_literature_mpi.py
@@ -293,7 +293,6 @@ EFB_Figure6_k2_params = {
     ],
 )
 def test_prb_82_085116_2010(p):
-
     COMM = MPI.COMM_WORLD
 
     gt = p["gt"]
