### 2023-02-06 a09a99b
reset padded_kw ext to 1000 to avoid padding issues with larger brigades
