### 2023-03-10 ed1b4a7
dispersion func added to Solver classes
