+++ README.rst
@@ -1,78 +1,0 @@
-
-GGCE
-====
-
-.. image:: https://github.com/x94carbone/GGCE/actions/workflows/ci.yml/badge.svg
-   :target: https://github.com/x94carbone/GGCE/actions/workflows/ci.yml
-
-.. image:: https://codecov.io/gh/x94carbone/GGCE/branch/master/graph/badge.svg?token=6Q7EUWBW6O
-   :target: https://codecov.io/gh/x94carbone/GGCE
-   :alt: codecov
-
-.. image:: https://app.codacy.com/project/badge/Grade/bdb53153835a49fa8921b28a519b2ead
-   :target: https://www.codacy.com/gh/x94carbone/GGCE/dashboard?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=x94carbone/GGCE&amp;utm_campaign=Badge_Grade
-   
-.. image:: https://joss.theoj.org/papers/688705844ea344353b86815d8345f8d5/status.svg
-   :target: https://joss.theoj.org/papers/688705844ea344353b86815d8345f8d5
-
-.. image:: https://zenodo.org/badge/DOI/10.48550/arXiv.2210.12260.svg
-   :target: https://doi.org/10.48550/arXiv.2210.12260
-
-Generalized Green's function Cluster Expansion
-
-Installation
-------------
-
-We provide detailed information on how to install the GGCE package in our `documentation <https://matthewcarbone.github.io/GGCE/installation.html>`__.
-
-Acknowledgements
-----------------
-
-.. inclusion-marker-acknowledgements-begin
-
-This software is based upon work supported by the U.S. Department of Energy, Office of Science, Office of Advanced Scientific Computing Research, Department of Energy Computational Science Graduate Fellowship under Award Number DE-FG02-97ER25308. This work also used theory and computational resources of the Center for Functional Nanomaterials, which is a U.S. Department of Energy Office of Science User Facility, and the Scientific Data and Computing Center, a component of the Computational Science Initiative, at Brookhaven National Laboratory under Contract No. DE-SC0012704. We also acknowledge support from the Natural Sciences and Engineering Research Council of Canada (NSERC) and the Stewart Blusson Quantum Matter Institute (SBQMI).
-
-.. inclusion-marker-acknowledgements-end
-
-
-Citation
---------
-
-If you use this software, please cite our preprint here:
-
-.. code-block::
-
-    @article{carbone2022generalized,
-      title={The Generalized Green's function Cluster Expansion: A Python package for simulating polarons},
-      author={Carbone, Matthew R and Fomichev, Stepan and Millis, Andrew J and Berciu, Mona and Reichman, David R and Sous, John},
-      journal={arXiv preprint arXiv:2210.12260},
-      year={2022}
-    }
-    
-We are also in the process of `editorial/peer review at the Journal of Open Source Software <https://github.com/openjournals/joss-reviews/issues/4884>`__.
-
-If you make references to the theory or application of the GGCE methodology, please consider citing the following two manuscripts: `Carbone, Reichman et al. <https://journals.aps.org/prb/abstract/10.1103/PhysRevB.104.035106>`__ and `Carbone, Millis et al <https://journals.aps.org/prb/abstract/10.1103/PhysRevB.104.L140307>`__:
-
-.. code-block::
-
-   @article{carbone2021,
-     title={Numerically exact generalized Green's function cluster expansions for electron-phonon problems},
-     author={Carbone, Matthew R and Reichman, David R and Sous, John},
-     journal={Phys. Rev. B.},
-     volume={104},
-     number={3},
-     pages={035106},
-     year={2021},
-     publisher={APS}
-   }
-
-   @article{carbone2021bond,
[…]
