+++ README.md
@@ -12,8 +12,7 @@
 
 **Numerically exact Green's functions for lattice polaron models, easily and efficiently**
 
-_If you use this work, please cite us!_
-   
+_If you use this work, please cite us!_ <br>
 [GGCE Software](https://doi.org/10.48550/arXiv.2210.12260) | [Original Paper](https://doi.org/10.1103/PhysRevB.104.035106) | [Bond-Peierls Paper](https://doi.org/10.1103/PhysRevB.104.L140307)
 
 </div>
