### 2023-04-27 0a30eb0
Update pyproject.toml
Attempt to debug build issues on master
