+++ pyproject.toml
@@ -31,6 +31,7 @@ classifiers = [
 ]
 dependencies = [
     "numpy<1.24",
+    'cython>=0.29.32',
     "scipy",
     "loguru",
     "monty",
