+++ pyproject.toml
@@ -31,7 +31,6 @@ classifiers = [
 ]
 dependencies = [
     "numpy<1.24",
-    'cython>=0.29.32',
     "scipy",
     "loguru",
     "monty",
