+++ .github/workflows/ci.yml
@@ -129,13 +129,13 @@ jobs:
         mpi: ${{ matrix.mpi }}
 
     - name: Install mpi4py and pytest-mpi
-      run: pip install "mpi4py==petsc-3.18.4" pytest-mpi
+      run: pip install mpi4py pytest-mpi
 
     - name: Run MPI tests
       run: mpiexec -n 2 coverage run --rcfile=ggce/_tests/mpi/setup_pytest_mpi.cfg -m pytest -v --with-mpi ggce/_tests/mpi/*.py
 
     - name: Setup PETSc
-      run: pip install petsc petsc4py
+      run: pip install "petsc==3.18.4" "petsc4py==3.18.4"
       env:
         PETSC_CONFIGURE_OPTIONS: "--with-scalar-type=complex --download-mumps --download-scalapack"
 
