+++ .github/workflows/ci.yml
@@ -103,7 +103,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
-        python-version: [3.7, 3.8, 3.9]
+        python-version: [3.7, 3.8, 3.9, "3.10", 3.11]
         mpi: ["openmpi"]
 
     steps:
@@ -134,23 +134,15 @@ jobs:
     - name: Run MPI tests
       run: mpiexec -n 2 coverage run --rcfile=ggce/_tests/mpi/setup_pytest_mpi.cfg -m pytest -v --with-mpi ggce/_tests/mpi/*.py
 
-    - name: Setup PETSc
-      run: pip install "petsc==3.18.4" "petsc4py==3.18.4"
-      env:
-        PETSC_CONFIGURE_OPTIONS: "--with-scalar-type=complex --download-mumps --download-scalapack"
+    # - name: Combine coverage reports from all ranks and generate xml report
+    #   run: coverage combine && coverage xml
 
-    - name: Run PETSc tests
-      run: mpiexec -n 4 coverage run --rcfile=ggce/_tests/petsc/setup_pytest_petsc_mpi.cfg -m pytest -v --with-mpi ggce/_tests/petsc/*.py
-
-    - name: Combine coverage reports from all ranks and generate xml report
-      run: coverage combine && coverage xml
-
-    - name: Upload code coverage
-      uses: codecov/codecov-action@v2
-      timeout-minutes: 10
-      with:
-        token: ${{ secrets.CODECOV_TOKEN }}
-        fail_ci_if_error: false
+    # - name: Upload code coverage
+    #   uses: codecov/codecov-action@v2
+    #   timeout-minutes: 10
+    #   with:
+    #     token: ${{ secrets.CODECOV_TOKEN }}
+    #     fail_ci_if_error: false
 
   build_and_publish:
     name: Build and publish to PyPI
