### 2023-05-04 6ffe26a
Update ci_petsc.yml
