### 2023-05-04 82de6cf
Remove Python v3.10
Test seemed to fail for v3.10 but might pass for others. Want to see if petsc and petsc4py successful compile now that pip is pinned to 23.0.1.
