### 2023-05-04 28a0cb3
Enable ```--use-pep517``` for petsc install
Suggestion from warning message: 

```DEPRECATION: petsc4py is being installed using the legacy 'setup.py install' method, because it does not have a 'pyproject.toml' and the 'wheel' package is not installed. pip 23.1 will enforce this behaviour change. A possible replacement is to enable the '--use-pep517' option. Discussion can be found at https://github.com/pypa/pip/issues/8559```
