### 2023-05-04 eaa253f
Add back python v3.10
Lets see if this one still fails to build petsc!
