+++ .github/workflows/ci_petsc.yml
@@ -56,7 +56,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
-        python-version: [3.7, 3.8, 3.9, 3.11]
+        python-version: [3.7, 3.8, 3.9, "3.10", 3.11]
         mpi: ["openmpi"]
 
     steps:
