### 2023-05-05 matthewcarbone/GGCE#77 @Neutrino155
Missing documentation
https://github.com/openjournals/joss-reviews/issues/5115

The documentation and tutorials look good so far but seem to be incomplete.

Some examples:

- There are some 'TK' comments in the Installation and Introduction sections.
- The Finite Temperature Calculation documentation (https://matthewcarbone.github.io/GGCE/tutorials/finite_t.html) seems to be missing although `ggce` appears to have this functionality.

Are there plans to complete these? I was also wondering if the Overview section is complete or whether it's just a placeholder for now.
