### 2023-05-05 9c8945d
Require Cython for building source C files
