+++ .github/workflows/ci_petsc.yml
@@ -89,7 +89,7 @@ jobs:
       run: mpiexec -n 2 coverage run --rcfile=ggce/_tests/mpi/setup_pytest_mpi.cfg -m pytest -v --with-mpi ggce/_tests/mpi/*.py
 
     - name: Setup PETSc
-      run: pip install "petsc==3.19.1" "petsc4py==3.19.1" --no-cache-dir --no-binary=petsc --no-binary=petsc4py
+      run: pip install "petsc==3.19.1" "petsc4py==3.19.1"/ --no-cache-dir --no-binary=petsc --no-binary=petsc4py
       env:
         PETSC_CONFIGURE_OPTIONS: "--with-scalar-type=complex --download-mumps --download-scalapack"
 
@@ -102,6 +102,6 @@ jobs:
     - name: Upload code coverage
       uses: codecov/codecov-action@v2
       timeout-minutes: 10
-      with:i
+      with:
         token: ${{ secrets.CODECOV_TOKEN }}
         fail_ci_if_error: false
