### 2023-05-05 cfbb53d
Doesn't like `:all:` so replaced with more verbose code
