### 2023-05-09 d9408ac
focus on py3.10 only, remove serial tests
