### 2023-05-09 b5e7b0b
only petsc, verbose pip
