+++ .github/workflows/ci_petsc.yml
@@ -82,11 +82,8 @@ jobs:
     - name: Install mpi4py and pytest-mpi
       run: pip install mpi4py pytest-mpi
 
-    - name: Run MPI tests
-      run: mpiexec -n 2 coverage run --rcfile=ggce/_tests/mpi/setup_pytest_mpi.cfg -m pytest -v --with-mpi ggce/_tests/mpi/*.py
-
     - name: Setup PETSc
-      run: pip install "petsc==3.19.1" "petsc4py==3.19.1" --no-cache-dir --no-binary=petsc --no-binary=petsc4py
+      run: pip install petsc --no-cache-dir --no-binary=petsc --verbose
       env:
         PETSC_CONFIGURE_OPTIONS: "--with-scalar-type=complex --download-mumps --download-scalapack"
 
