### 2023-05-09 7cdef3e
install setuptools manually
