+++ .github/workflows/ci_petsc.yml
@@ -57,7 +57,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest]
-        python-version: ["3.10"]
+        python-version: ["3.9", "3.10"]
         mpi: ["openmpi"]
 
     steps:
