+++ .github/workflows/ci_petsc.yml
@@ -57,7 +57,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest]
-        python-version: ["3.9", "3.10"]
+        python-version: ["3.10"]
         mpi: ["openmpi"]
 
     steps:
@@ -85,20 +85,10 @@ jobs:
     - name: Install wheel and setuptools manually
       run: pip install wheel setuptools
 
+    - name: Downgrade pip
+      run: pip install pip==23.0.1
+
     - name: Setup PETSc
       run: pip install petsc --verbose
       env:
         PETSC_CONFIGURE_OPTIONS: "--with-scalar-type=complex --download-mumps --download-scalapack"
-
-    - name: Run PETSc tests
-      run: mpiexec -n 4 coverage run --rcfile=ggce/_tests/petsc/setup_pytest_petsc_mpi.cfg -m pytest -v --with-mpi ggce/_tests/petsc/*.py
-
-    - name: Combine coverage reports from all ranks and generate xml report
-      run: coverage combine && coverage xml
-
-    - name: Upload code coverage
-      uses: codecov/codecov-action@v2
-      timeout-minutes: 10
-      with:
-        token: ${{ secrets.CODECOV_TOKEN }}
-        fail_ci_if_error: false
