### 2023-05-12 c0ddd7c
uninstall and reinstall pip using ensurepip module
