### 2023-05-12 8a33660
push -y to avoid prompt with uninstall
