+++ .github/workflows/ci_petsc.yml
@@ -72,6 +72,7 @@ jobs:
     - name: Install requirements
       run: |
         pip -V
+        pip3 -V
         bash scripts/install.sh
         bash scripts/install.sh test
 
@@ -85,9 +86,12 @@ jobs:
 
     - name: Downgrade pip
       run: |
-        pip uninstall pip
+        pip -V
+        pip3 -V
+        pip uninstall -y pip
         python -m ensurepip
         pip -V
+        pip3 -V
 
     - name: Setup PETSc
       run: |
