### 2023-06-09 4692228
Small fixes in paper.md
