+++ paper.md
@@ -67,7 +67,7 @@ Research on polarons has been divided into two thrusts: fundamental theoretical 
 
 The GGCE method is a _numerically exact_ extension of a family of variational approaches known in the theoretical physics community as Momentum Average (MA) methods [@Berciu2006prl; @Goodvin2006prb]. Details on the theoretical framework of GGCE can be found in Carbone _et al_ [@carbone2021numerically; @carbone2021bond]. Our code, named for the method, is a Python package meant to make implementing the GGCE framework as straightforward as possible. In addition, through only slight modifications to our standard API, the user can invoke powerful PETSc sparse solvers for massively parallel computations at scale [@petsc-web-page; @petsc-user-ref; @petsc-efficient; @PetscSF2022]. 
 
-A fundamental insight of the MA approximation is to utilize a variational space formed of clouds of spatially clustered phonon configurations with the electron allowed to be anywhere in the system. Through comparison with exact methods, this approximation was shown to yield quantitatively accurate results. In order to systematically converge MA to the limit of infinite Hilbert space dimension, the cloud size and total phonon number, which serve as control parameters, are taken to infinity [@Goodvin2006prb]. This, however, requires derivation of a set of equations corresponding to a given cloud size for all cloud sizes smaller than a cutoff. This cutoff is then increased until convergence is achieved. The ever-increasing complexity of the structure of the system of equations at large cloud sizes means that this approach can very quickly become intractable to do by hand [@Dominic; @Sous1; @Sous2], especially in the regime of small phonon energies where large cloud sizes are usually needed in order to converge to the numerically exact limit. Carbone _et al_ proposed a a generalized implementation of the MA method which automates the generation and solution of the systems of equations for arbitrary cloud sizes [@carbone2021numerically]. Benchmarks of GGCE on several model systems verified that convergence with cloud size is fast, rendering this an efficient and controlled numerically exact method even in challenging parameter regimes. 
+A fundamental insight of the MA approximation is to utilize a variational space formed of clouds of spatially clustered phonon configurations with the electron allowed to be anywhere in the system. Through comparison with exact methods, this approximation was shown to yield quantitatively accurate results. In order to systematically converge MA to the limit of infinite Hilbert space dimension, the cloud size and total phonon number, which serve as control parameters, are taken to infinity [@Goodvin2006prb]. This, however, requires the derivation of a set of equations corresponding to a given cloud size for all cloud sizes smaller than a cutoff. This cutoff is then increased until convergence is achieved. The ever-increasing complexity of the structure of the system of equations at large cloud sizes means that this approach can very quickly become intractable to do by hand [@Dominic; @Sous1; @Sous2], especially in the regime of small phonon energies where large cloud sizes are usually needed in order to converge to the numerically exact limit. Carbone _et al_ proposed a generalized implementation of the MA method which automates the generation and solution of the systems of equations for arbitrary cloud sizes [@carbone2021numerically]. Benchmarks of GGCE on several model systems verified that convergence with cloud size is fast, rendering this an efficient and controlled numerically exact method even in challenging parameter regimes. 
 
 Previous studies using GGCE focused on polarons at zero temperature. We also include a new functionality which allows the study of polarons at finite temperature. Here, we make use of the Thermofield Double formalism [@suzuki1985thermo; @takahashi1996thermo], which exactly maps any given model at finite temperature to one at zero temperature with couplings to real and fictitious phonons. This model can be solved naturally using the apparatus of the zero-temperature GGCE method. Benchmarks of this approach are ongoing, and preliminary results suggest that the method may be competitive with state-of-the-art methods. A paper with these results is currently in preparation.
 
@@ -85,15 +85,15 @@ The choice of model completely defines the type of electron-phonon coupling used
 
 ## Systems
 
[…]
