### 2023-07-27 #5115 @jarvist
@matthewcarbone , have you had time to consider the tweaks to the paper?
There's just a couple of check pointson the paper & @st3r4g 's issue.

@Neutrino155  - what is required for you to sign off on the review? Currently there's just a couple of check boxes for the paper, and one for automated tests.
