### 2023-07-28 #5115 @jarvist
I must admit I have found GitHub notifications to have increasingly lower SNR, including for my editorial duties!

I think there were some issues with versions of pip in terms of being able to 'clean install' and run the tests - is that fixed now?

@st3r4g and @Neutrino155 - if you could check & read the paper, and hopefully we'll be able to move this forwards.
