### 2023-07-28 #5115 @matthewcarbone
> I think there were some issues with versions of pip in terms of being able to 'clean install' and run the tests - is that fixed now?

No not yet. And unfortunately we're having some trouble with the PETSc build for some reason (any help on that front is appreciated!). I'll try to get to it all.
