### 2023-08-29 #5115 @Neutrino155
@jarvist @matthewcarbone

I have completed my review. I have a couple of comments:

- I think the package would still benefit from some documentation on the finite temperature implementation of the code. I opened an issue here: https://github.com/matthewcarbone/GGCE/issues/77. Nonetheless, the current documentation still covers the core functionality of the package so I'm happy to tick that off.

- I think the PETSc still doesn't build: https://github.com/matthewcarbone/GGCE/pull/78, but that seems to be a separate issue that doesn't effect the core functionality of the package and can be fixed at a later date.

Overall, I'm able to tick everything off on the checklist :)
