### 2023-08-29 #5115 @jarvist
OK @matthewcarbone  - I'm happy to recommend this for publication. The above checklist is for us to work through together.

Generating the software release + DOI is quite an important step, that can take a bit of time and space. (I usually use Zenodo to make this directly from a GitHub 'Tag'.)

Congratulations!
