### 2023-08-29 #5115 @matthewcarbone
@Neutrino155 thanks so much for the review. Yes unfortunately we are still struggling with the PETSc build... This was one of those situations where it seemed to work one day and the next day it just stopped working. We'll try to get it up and running but as you say it does not impact the core functionality of GGCE. I think this is actually a petsc4py packaging problem, or at least it seems like it.

@jarvist I'll handle the DOIs and whatnot, and all of the author tasks above (copying/pasting here so I can check things off as I complete them). I'll be using Zenodo to checkpoint the release. 👍

> ## Additional Author Tasks After Review is Complete
> * [ ]  Double check authors and affiliations (including ORCIDs)
> * [ ]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * [ ]  Archive the release on Zenodo/figshare/etc and post the DOI here.
> * [ ]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [ ]  Make sure that the license listed for the archive is the same as the software license.
