+++ paper.bib
@@ -6,7 +6,8 @@
   number={3},
   pages={035106},
   year={2021},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevB.104.035106},
 }
 
 @article{carbone2021bond,
@@ -17,7 +18,8 @@
   number={14},
   pages={L140307},
   year={2021},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevB.104.L140307}
 }
 
 @book{ephReview1,
@@ -25,14 +27,16 @@
   author={Alexandrov, Alexandre S and Devreese, Jozef T},
   volume={159},
   year={2010},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-3-642-01896-1}
 }
 
 @book{ephReview2,
   title={Fundamentals of semiconductors: physics and materials properties},
   author={Peter, Yu and Cardona, Manuel},
   year={2010},
-  publisher={Springer Science \& Business Media}
+  publisher={Springer Science \& Business Media},
+  doi={10.1080/00107514.2012.661781}
 }
 
 @article{solarcells,
@@ -43,7 +47,8 @@
   number={6},
   pages={2162--2173},
   year={2021},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acsenergylett.1c00506}
 }
 
 @article{semiconductors,
@@ -54,7 +59,8 @@
   number={5},
   pages={491--502},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41563-020-0647-2}
 }
 
 @incollection{Mahan,
@@ -62,7 +68,8 @@
   author={Mahan, Gerald D},
   booktitle={Condensed Matter in a Nutshell},
   year={2010},
-  publisher={Princeton University Press}
+  publisher={Princeton University Press},
+  doi={10.1515/9781400837021}
 }
 
 @article{MA,
@@ -80,17 +87,6 @@
   url = {https://link.aps.org/doi/10.1103/PhysRevLett.97.036402}
 }
 
-@article{Berciu2006prl,
-  title={Green’s function of a dressed particle},
-  author={Berciu, Mona},
-  journal={Physical review letters},
-  volume={97},
-  number={3},
-  pages={036402},
-  year={2006},
-  publisher={APS}
-}
-
 @article{Goodvin2006prb,
   title={Green’s function of the Holstein polaron},
   author={Goodvin, Glen L and Berciu, Mona and Sawatzky, George A},
@@ -99,7 +95,8 @@
   number={24},
   pages={245104},
   year={2006},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevB.74.245104}
 }
 
 @article{Sous1,
@@ -110,7 +107,8 @@
   number={1},
   pages={1--7},
   year={2017},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41598-017-01228-y}
 }
 
 @article{Sous2,
@@ -121,7 +119,8 @@
   number={24},
   pages={247001},
   year={2018},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevLett.121.247001}
 }
 
 @article{Dominic,
@@ -132,7 +131,8 @@
   number={26},
   pages={266605},
   year={2010},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevLett.105.266605}
 }
 
 @article{giustino2017electron,
[…]
