### 2023-09-17 2625f15
Update README.md
Minor updates to the README
