### 2023-09-29 #5115 @jarvist
OK, that's me @matthewcarbone ! Congratulations on getting this far. The 'Editor in Chief' team should be along shortly for the final checks and balances.

EiC; the paper has a disclaimer from the US Govt in the acknowledgments. It all looks fairly benign & boiler-plate, but I've never seen one of these in a JOSS I've edited, so I would like to flag it up for your attention.
